Moved modification types into their own classes and files
