+++ app/src/main/scala/environment/environment-builder/Modifications.scala
@@ -1,0 +1,9 @@
+package environment.modifications
+
+import environment.layer._
+
+
+// Terrain Modifications
+trait TerrainModification {
+  def modify(layer: Layer, constructionLayer: ConstructionLayer): Unit
+}
